requirements.txt
requests==2.32.3
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.1.3
websocket-client==1.8.0
streamlit==1.40.1
plotly==5.24.1
